clean up old tests, move flattener files into sub package
