+++ projects/batfish/src/main/java/org/batfish/grammar/juniper/JuniperFlattener.java
@@ -3,8 +3,8 @@ package org.batfish.grammar.juniper;
 import com.google.common.collect.ImmutableList;
 import java.util.ArrayList;
 import java.util.List;
-import org.batfish.grammar.Flattener;
-import org.batfish.grammar.FlattenerLineMap;
+import org.batfish.grammar.flattener.Flattener;
+import org.batfish.grammar.flattener.FlattenerLineMap;
 import org.batfish.grammar.juniper.JuniperParser.Bracketed_clauseContext;
 import org.batfish.grammar.juniper.JuniperParser.Juniper_configurationContext;
 import org.batfish.grammar.juniper.JuniperParser.StatementContext;
