+++ projects/batfish/src/test/java/org/batfish/main/BatfishTest.java
@@ -30,9 +30,7 @@ import java.util.SortedSet;
 import java.util.TreeMap;
 import java.util.TreeSet;
 import org.batfish.common.BatfishException;
-import org.batfish.common.BatfishLogger;
 import org.batfish.common.util.CommonUtil;
-import org.batfish.config.Settings;
 import org.batfish.datamodel.Configuration;
 import org.batfish.datamodel.ConfigurationFormat;
 import org.batfish.datamodel.Edge;
@@ -42,9 +40,6 @@ import org.batfish.datamodel.answers.Answer;
 import org.batfish.datamodel.answers.AnswerStatus;
 import org.batfish.datamodel.answers.ParseStatus;
 import org.batfish.datamodel.answers.ParseVendorConfigurationAnswerElement;
-import org.batfish.grammar.Flattener;
-import org.batfish.grammar.FlattenerLineMap;
-import org.batfish.grammar.VendorConfigurationFormatDetector;
 import org.batfish.representation.host.HostConfiguration;
 import org.batfish.vendor.VendorConfiguration;
 import org.junit.Rule;
@@ -54,7 +49,6 @@ import org.junit.rules.TemporaryFolder;
 
 /** Tests for {@link Batfish}. */
 public class BatfishTest {
-  private static final String PAN_TESTCONFIGS_PREFIX = "org/batfish/grammar/palo_alto/testconfigs/";
 
   @Rule public TemporaryFolder _folder = new TemporaryFolder();
 
@@ -93,46 +87,6 @@ public class BatfishTest {
     assertThat(
         answer.getAnswerElements().get(0).prettyPrint(),
         containsString("Could not parse question"));
-  }
-
-  @Test
-  public void testFlattenConfig() throws IOException {
-    String nestedConfig = "nested-config-line-tracking";
-    Flattener flattener =
-        Batfish.flatten(
-            CommonUtil.readResource(PAN_TESTCONFIGS_PREFIX + nestedConfig),
-            new BatfishLogger(BatfishLogger.LEVELSTR_OUTPUT, false),
-            new Settings(),
-            ConfigurationFormat.PALO_ALTO_NESTED,
-            VendorConfigurationFormatDetector.BATFISH_FLATTENED_PALO_ALTO_HEADER);
-    FlattenerLineMap lineMap = flattener.getOriginalLineMap();
-
-    // Confirm looking up different characters results in the correct original line numbers
-    // Lookup deviceconfig
-    assertThat(lineMap.getOriginalLine(2, 4), equalTo(2));
-    // Lookup nested-config
-    assertThat(lineMap.getOriginalLine(2, 33), equalTo(4));
-    // Lookup blah
-    assertThat(lineMap.getOriginalLine(3, 24), equalTo(5));
-    // Lookup 1.1.1.1
-    assertThat(lineMap.getOriginalLine(4, 52), equalTo(8));
-    // Lookup stuff
-    assertThat(lineMap.getOriginalLine(6, 11), equalTo(15));
-    // Lookup bar
-    assertThat(lineMap.getOriginalLine(6, 21), equalTo(16));
-    // Lookup stuff
-    assertThat(lineMap.getOriginalLine(7, 11), equalTo(15));
-    // Lookup baz
-    assertThat(lineMap.getOriginalLine(7, 21), equalTo(16));
-
-    // Confirm looking up a character not corresponding to an original line results in the last
-    // original line number for that constructed line
-    // Lookup 's' in set
-    assertThat(lineMap.getOriginalLine(2, 0), equalTo(4));
-    // Lookup 'e' in set
-    assertThat(lineMap.getOriginalLine(5, 1), equalTo(9));
-    // Lookup 't' in set
-    assertThat(lineMap.getOriginalLine(6, 2), equalTo(16));
   }
 
   @Test
