--- projects/batfish/src/test/resources/org/batfish/grammar/palo_alto/testconfigs/nested-config-line-tracking
@@ -1,18 +1,0 @@
-# comment
-deviceconfig {
-    system {
-        hostname nested-config;
-        blah;
-        dns-setting {
-            servers {
-                primary 1.1.1.1;
-                secondary 2.2.2.2;
-            }
-        }
-    }
-}
-shared {
-    stuff {
-        foo [bar baz];
-    }
-}
